Gas optimisations (#470)
* Gas optimisations

* Unused import

+++ contracts/implementation/KeeperRewards.sol
@@ -141,7 +141,7 @@ contract KeeperRewards is IKeeperRewards {
             /* safe due to explicit bounds check for settlementTokenPrice above */
             /* (wei * Settlement / ETH) / fixed point (10^18) = amount in settlement */
             bytes16 weiSpent = ABDKMathQuad.fromUInt(_gasPrice * gasUsed);
-            bytes16 settlementTokenPrice = ABDKMathQuad.fromUInt(uint256(_settlementTokenPrice));
+            bytes16 settlementTokenPrice = ABDKMathQuad.fromUInt(_settlementTokenPrice);
             return ABDKMathQuad.toUInt(ABDKMathQuad.div(ABDKMathQuad.mul(weiSpent, settlementTokenPrice), FIXED_POINT));
         }
     }
